Added Alert Dialog

+++ app/src/main/java/com/umutcansahin/todoapp/ui/home/HomeFragment.kt
@@ -1,5 +1,7 @@
 package com.umutcansahin.todoapp.ui.home
 
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
@@ -7,6 +9,7 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.navigation.Navigation
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.umutcansahin.todoapp.R
 import com.umutcansahin.todoapp.databinding.FragmentHomeBinding
 import com.umutcansahin.todoapp.ui.home.adapter.ToDoListAdapter
@@ -38,16 +41,28 @@ class HomeFragment : Fragment(R.layout.fragment_home) {
         binding.recyclerView.layoutManager = LinearLayoutManager(requireContext())
 
         binding.floatingActionButton.setOnClickListener {
-            val action = HomeFragmentDirections.actionNavigationHomeToAddFragment(isInsert = true,id = 0)
+            val action =
+                HomeFragmentDirections.actionNavigationHomeToAddFragment(isInsert = true, id = 0)
             Navigation.findNavController(it).navigate(action)
         }
         toDoListAdapter.onItemClick = {
-            viewModel.deleteToDo(it)
+
+            val alert = MaterialAlertDialogBuilder(requireContext())
+            alert.setTitle(R.string.alert)
+            alert.setIcon(R.drawable.ic_baseline_warning_amber_24)
+            alert.setMessage(R.string.alert_mesaj)
+            alert.setPositiveButton(R.string.yes) { _, _ ->
+                viewModel.deleteToDo(it)
+            }
+            alert.setNegativeButton(R.string.no) { _, _ ->
+
+            }
+            alert.show()
         }
     }
 
     private fun observeEvent() {
-        viewModel.uiState.observe(viewLifecycleOwner, Observer { todo->
+        viewModel.uiState.observe(viewLifecycleOwner, Observer { todo ->
             todo?.let {
                 toDoListAdapter.updateList(it)
             }
